feat: moved keycloak token logic
